@@ -1,22 +1,23 @@
 plugins {
-    id("java")
+    id("com.android.library")
 }
 
-sourceCompatibility = 1.7
-targetCompatibility = 1.7
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
 
 dependencies {
     implementation project(':rx-extensions')
     implementation 'javax.inject:javax.inject:1'
-    implementation 'com.squareup.retrofit2:retrofit:2.1.0'
-    implementation 'com.squareup.retrofit2:adapter-rxjava:2.1.0'
-    implementation 'com.squareup.retrofit2:converter-gson:2.1.0'
-    implementation 'com.squareup.okhttp3:okhttp:3.4.1'
-    implementation "com.google.code.gson:gson:2.4"
+    api 'com.squareup.retrofit2:retrofit:2.1.0'
+    api 'com.squareup.retrofit2:adapter-rxjava:2.1.0'
+    api 'com.squareup.retrofit2:converter-gson:2.1.0'
+    api 'com.squareup.okhttp3:okhttp:3.4.1'
+    api "com.google.code.gson:gson:2.4"
     implementation "org.json:json:20131018"
+    implementation 'com.google.code.findbugs:jsr305:3.0.2'
 
     // Recycler view helpers
-    implementation 'com.github.jacek-marchwicki.recyclerview-changes-detector:universal-adapter-java:1.0.0'
+    implementation("com.github.jacek-marchwicki.recyclerview-changes-detector:universal-adapter:2.0.1")
 
     implementation "io.reactivex:rxjava:1.3.8"
     implementation "javax.annotation:javax.annotation-api:1.2"
@@ -27,7 +28,21 @@ dependencies {
     testImplementation 'com.google.truth:truth:0.42'
 }
 
-test {
-    include '**/*Test.class'
-    maxParallelForks 1
+android {
+    compileSdkVersion 28
+
+    defaultConfig {
+        minSdkVersion 14
+        targetSdkVersion 28
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
+    }
+
+    packagingOptions {
+        exclude 'LICENSE.txt'
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
