@@ -4,8 +4,8 @@ apply plugin: 'maven'
 group='com.github.jacek-marchwicki.rx-java-extensions'
 // When changing update README file
 version='1.0.4'
-sourceCompatibility = 1.7
-targetCompatibility = 1.7
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
 
 dependencies {
     implementation "io.reactivex:rxjava:1.3.8"
