@@ -18,7 +18,7 @@
 package com.appunite.rx.android.widget;
 
 import android.app.Activity;
-import android.support.v4.app.ActivityCompat;
+import androidx.core.app.ActivityCompat;
 
 import javax.annotation.Nonnull;
 
