@@ -16,7 +16,7 @@
 
 package com.appunite.rx.android.widget;
 
-import android.support.v4.view.ViewCompat;
+import androidx.core.view.ViewCompat;
 import android.view.View;
 import android.view.ViewTreeObserver;
 
